Rename `datasource::datasource` to `datasource::provider`.
